fix(dataflows): the vendor-routing layer truncated the recent rows too
Found by running the real market analyst end to end against AMD on the
already-fixed tool wrapper: the model still received rows ending 2026-06-26
for a 2026-08-31 analysis.

dataflows.interface.route_to_vendor calls _truncate on EVERY vendor return,
below the agent tool wrappers, and it kept the head unconditionally. For the
OHLCV vendors — which emit ascending rows — the recent data was therefore
already gone before any agent saw it, and truncating again in the wrapper
could not recover it. Fixing only core_stock_tools was cosmetic.

Route _truncate through truncate_series_text as well. Prose (news,
fundamentals) still head-truncates, so the lede is unchanged. Character
budgets are untouched in both layers.

Move truncate_text / truncate_series_text into dataflows/text_truncation.py, a
leaf module, so both layers share one implementation: agent_utils imports the
tool modules, which import dataflows.interface, so keeping the helpers in
agent_utils and importing them from interface would be circular. agent_utils
re-exports them, so every existing agent import keeps working.

Also report the older-row sampling as "1 row in N" rather than an ungrammatical
"every 1th row" when the stride degenerates to 1.
